requirements.txt
fastapi
uvicorn[standard]
openenv-core
pydantic>=2.0
requests
openai